+++ StepTrackerTCA/Features/Workout/Creator/WorkoutCreator/Feature/WorkoutCreatorFeature.swift
@@ -44,10 +44,19 @@ struct WorkoutCreatorFeature {
                 return .none
                 
             case let .warmupGoalChanged(goal):
-                //state.warmupGoal = goal
+                state.warmupGoal = goal
+                return .none
+                
+            case let .warmupTimeChange(time):
+                state.warmUpGoalTime = time
                 return .none
                 
             case let .coolDownGoalChanged(goal):
+                state.coolDownGoal = goal
+                return .none
+                
+            case let .coolDownTimeChange(time):
+                state.coolDownTime = time
                 return .none
                 
                 // MARK: - View Actions
@@ -77,137 +86,16 @@ struct WorkoutCreatorFeature {
                 
                 // MARK: - Child
             case let .destination(.presented(.openWodCreator(.delegate(.wodCreated(workout))))):
-//            case let .wodCreator(.delegate(.wodCreated( workout))):
                 print("Otrzymano workout:")
                 dump(workout)
                 return .none
                 
             case .destination:
                 return .none
-                
-//            default:
-//                return .none
             }
             
         }
         .ifLet(\.$destination, action: \.destination)
-//        Scope(state: \.wodCreator, action: \.wodCreator) {
-//            WodCreatorFeature()
-//        }
-    }
-}
-
-// MARK: - Action
-
-/// Implementation of `WorkoutCreatorFeature` action
-extension WorkoutCreatorFeature {
-    
-    @CasePathable
-    enum Action: ViewAction, BindableAction {
-        
-        // MARK: - Binding Action
-        
-        /// Handles changes in bindings for the state.
-        case binding(BindingAction<State>)
-        
-        // MARK: - Actions
-        
-        case selectedWorkoutActivityPickerChange(WorkoutActivityType)
-        
-        case selectedWorkoutLocationPickerChange(WorkoutLocationType)
-        
-        case workoutTitleChanged(String)
-        
-        case warmupGoalChanged(SimpleWorkoutGoal)
-        
-        case coolDownGoalChanged(SimpleWorkoutGoal)
-        
-        // MARK: - View actions
-        
-        /// Used for view actions.
-        case view(View)
-        
-        enum View {
-            
-            /// The action responsible for completing tasks as soon as the view is displayed.
-            case viewDidAppear
-            
-            case cancelButtonTapped
-            
-            case workoutTitleSheetTapped
-            
-            case wodSheetTapped
-            
-            case workoutTitleSheetDismissed
-        }
-        
-        // MARK: - Destination
-        
-        case destination(PresentationAction<Destination.Action>)
-        
-        // MARK: - Child
-        
-//        case wodCreator(WodCreatorFeature.Action)
-        
     }
     
 }
-
-// MARK: - State
-
-/// Implementation of `WorkoutCreatorFeature` state
-extension WorkoutCreatorFeature {
-    
-    @ObservableState
-    struct State {
-        
-        var isWorkoutTitleSheetPresented: Bool = false
-        
-        var workoutTitle: String = ""
-        
-        var workoutActivityType: WorkoutActivityType = .crossTraining
-        
-        var availableWorkoutTypes: [WorkoutActivityType] = [.crossTraining]
-        
-        var workoutLocationType: WorkoutLocationType = .indoor
-        
-        var warmupGoal: SimpleWorkoutGoal = .open
-        
-        var coolDownGoal: SimpleWorkoutGoal = .open
-        
-        // MARK: - Destination
-        
-        @Presents var destination: Destination.State?
-        
-        // MARK: - Child
-        
-//        var wodCreator = WodCreatorFeature.State()
-        
-    }
-    
-}
-
-/// Implementation of `WorkoutCreatorFeature` destination
-extension WorkoutCreatorFeature {
-    
-    @Reducer
-    enum Destination {
-        
-        /// Represents the destination for displaying in `WodCreatorFeature`.
-        case openWodCreator(WodCreatorFeature)
-    }
-    
-}
-
-enum SimpleWorkoutGoal: CaseIterable, Hashable {
-    
-    case open
-    
-    var title: String {
-        switch self {
-        case .open: return "Open"
-        }
-    }
-    
-}
-
