+++ TASKS.md
@@ -427,4 +427,4 @@
 
 ### IOS-00045 Workout Creator sheet
     A: Implement a screen to manually add workouts.
-
+    B: Working on Creator
